fix: circular dependency
