+++ src/services/analytics/events/createLoadSafe.ts
@@ -1,4 +1,4 @@
-import { GTM_EVENT } from '@/services/analytics/analytics'
+import { GTM_EVENT } from '@/services/analytics/types'
 
 const CREATE_SAFE_CATEGORY = 'create-safe'
