+++ src/services/analytics/events/overview.ts
@@ -1,4 +1,4 @@
-import { GTM_EVENT } from '@/services/analytics/analytics'
+import { GTM_EVENT } from '@/services/analytics/types'
 
 const OVERVIEW_CATEGORY = 'overview'
