+++ src/services/analytics/events/txList.ts
@@ -1,4 +1,4 @@
-import { GTM_EVENT } from '@/services/analytics/analytics'
+import { GTM_EVENT } from '@/services/analytics/types'
 
 const TX_LIST_CATEGORY = 'tx-list'
