+++ src/services/analytics/events/wallet.ts
@@ -1,4 +1,4 @@
-import { GTM_EVENT } from '@/services/analytics/analytics'
+import { GTM_EVENT } from '@/services/analytics/types'
 
 const WALLET_CATEGORY = 'wallet'
