+++ src/services/analytics/types.ts
@@ -1,0 +1,8 @@
+export enum GTM_EVENT {
+  PAGEVIEW = 'pageview',
+  CLICK = 'customClick',
+  META = 'metadata',
+  SAFE_APP = 'safeApp',
+}
+
+export type EventLabel = string | number | boolean | null
